+++ cmd/download/main.go
@@ -17,7 +17,7 @@ import (
 	"github.com/bborbe/service"
 	"github.com/golang/glog"
 
-	libs3 "github.com/bborbe/s3-utils"
+	"github.com/bborbe/s3-utils"
 )
 
 func main() {
@@ -41,7 +41,7 @@ func (a *application) Run(
 	ctx context.Context,
 	sentryClient libsentry.Client,
 ) error {
-	s3Client := libs3.CreateS3Client(a.S3Url, a.S3AccessKey, a.S3SecretKey)
+	s3Client := s3utils.CreateS3Client(s3utils.URL(a.S3Url), s3utils.AccessKey(a.S3AccessKey), s3utils.SecretKey(a.S3SecretKey))
 	getObject, err := s3Client.GetObject(ctx, &s3.GetObjectInput{
 		Bucket: aws.String(a.BucketName),
 		Key:    aws.String(a.KeyName),