+++ cmd/upload/main.go
@@ -18,7 +18,7 @@ import (
 	"github.com/bborbe/service"
 	"github.com/golang/glog"
 
-	libs3 "github.com/bborbe/s3-utils"
+	"github.com/bborbe/s3-utils"
 )
 
 func main() {
@@ -42,7 +42,7 @@ func (a *application) Run(
 	ctx context.Context,
 	sentryClient libsentry.Client,
 ) error {
-	s3Client := libs3.CreateS3Client(a.S3Url, a.S3AccessKey, a.S3SecretKey)
+	s3Client := s3utils.CreateS3Client(s3utils.URL(a.S3Url), s3utils.AccessKey(a.S3AccessKey), s3utils.SecretKey(a.S3SecretKey))
 	uploader := manager.NewUploader(s3Client, func(u *manager.Uploader) {
 		u.PartSize = 5 * 1024 * 1024
 		u.Concurrency = 1