+++ s3-client-creator.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package s3
+package s3utils
 
 import (
 	"log"
@@ -15,22 +15,40 @@ import (
 	"github.com/golang/glog"
 )
 
+type URL string
+
+func (u URL) String() string {
+	return string(u)
+}
+
+type AccessKey string
+
+func (a AccessKey) String() string {
+	return string(a)
+}
+
+type SecretKey string
+
+func (s SecretKey) String() string {
+	return string(s)
+}
+
 func CreateS3Client(
-	s3Url string,
-	s3AccessKey string,
-	s3SecretKey string,
+	s3Url URL,
+	s3AccessKey AccessKey,
+	s3SecretKey SecretKey,
 ) *s3.Client {
 	return s3.New(
 		s3.Options{
 			UsePathStyle:     true,
-			EndpointResolver: s3.EndpointResolverFromURL(s3Url),
+			EndpointResolver: s3.EndpointResolverFromURL(s3Url.String()),
 			EndpointOptions: s3.EndpointResolverOptions{
-				DisableHTTPS: !strings.HasPrefix("https://", s3Url),
+				DisableHTTPS: !strings.HasPrefix("https://", s3Url.String()),
 			},
 			Credentials: aws.NewCredentialsCache(
 				credentials.NewStaticCredentialsProvider(
-					s3AccessKey,
-					s3SecretKey,
+					s3AccessKey.String(),
+					s3SecretKey.String(),
 					"",
 				),
 			),